Refactored peer id to long, according to peersim.Node. (re)Implemented a few Linkable interface related functions.

+++ src/buddycast/BuddyCastMessage.java
@@ -24,7 +24,7 @@ public class BuddyCastMessage {
     /**
      * The list of random peers
      */
-    Hashtable<Integer, Long> randomPeers;
+    Hashtable<Long, Long> randomPeers;
     /**
      * True, if the client is connectible;
      */
